selection: pretty: compact syntax printer as SelectionSYM
Summary:
the key point of this diff is to replace (and remove) `to_compact_syntax` by a type `struct SelectionCompact(String)` implementing `SelectionSYM`.

some corrections were made to the pretty printer in the process (see `first`, `false` and `label`).

there are then a bunch of simple refactoring that are really just about cutting down on repetition however i took the chance to upgrade a bunch of `hyperactor_mesh` testing of `sel!`.

Reviewed By: mariusae

Differential Revision: D74186930

fbshipit-source-id: e2038d244eb2497e6a0e147d837303516f090642