Add error reason and stack trace to reply files
Summary:
Unfortunately we cannot use the recommended reply file generation library: https://www.internalfb.com/wiki/PyPer/Dev/ErrorReporting/ due to OSS so we must implement what we need ourselves.

On the python side we write the error reason and stack trace to the reply file path but this seems to get intercepted by conda/mast/reply_file.py https://fburl.com/code/i3cvqh0v which writes to the reply file but doesn't show up in the mast ui. However, back in rust land we are able to read the file and then rewrite the contents and this somehow is able to satisfy mast with displaying it in the UI

To get this to work in a foolproof way, we wrap the training script in another script and launch the training script in a try block. When there is an uncaught exception in the training script, we will catch it and write a reply file. Using buck_resources to resolve a resource in an fbpkg in tupperware doesn't seem to work but if we include the buck_filegroup as one of the path actions in the fbpkg.builder the resource is bundled into the fbpkg and we can point to it. See https://fburl.com/code/a0d2hpbc for an example of someone else using this workaround

Reviewed By: highker

Differential Revision: D73817463

fbshipit-source-id: 36f73e5157e4f14bb302d80a4adf660bf1564187