Add mocked allocator for local testing
Summary:
Add a mew `MockMast` and corresponding Python binding to simulate Mast allocator stack using local or process allocators. Used mainly for testing scenarios such as task/task group failures and so on.

Also in this diff, introduce `PyAlloc` which serves as a generic class to represent an Alloc instead of previously hard-coding concrete types.

Reviewed By: highker

Differential Revision: D74257656

fbshipit-source-id: c78caa1681fe1ef642a541d869f4337f5e49dd95