make post infallible
Summary:
The current post is renamed "try_post", and we provide "post" as now being both non-async and infallible. The expectation is that the user monitors the channel for failure.

We should do some follow-on refactoring here, too, where "try_post" is also infallible, but returns a SendError<M> on its return channel -- capturing all possible errors. This would avoid having to maintain additional error handling paths in the caller.

Reviewed By: shayne-fletcher

Differential Revision: D74272868

fbshipit-source-id: 857b3ab4e379ff87e57e6ab2c8fbc47f865475c2