create python-level allocator
Summary:
In the process of making allocation/spawning expose a Future-based API, we would like to add Python methods to LocalAllocator and PythonAllocator.

In order to facilitate that, this diff makes the Rust types into `*AllocatorBase`, which are inherited by a Python class that exposes the actual public interface. Later we will add Python methods.

Also change `LocalAllocator.allocate` from a class method to an instance method, for more consistency between APIs.

Reviewed By: colin2328, mariusae

Differential Revision: D74262668

fbshipit-source-id: a589e233ccaf12d2d9e91cbcc9d0bd1d5dd57a6d