Refactor to call `.handle()` only once per message
Summary:
These does some minor refactoring to routing frames resolution, such
that instead of using `.handle()` to recurse on routing decisions within
a given node, we use a separate function to fully resolve to all routing
frames destined for peer-ranks and whether to deliver to the current rank.

This is meant to simplify a few components of routing for implementing
in-order delivery.  In particular, depending on the selection we may generate
*multiple* forward-messages from rank N to rank M for the same wrapped
message.  In this case, we need to know to wait for *both* of these so that
we can properly maintain the last-seq map we maintain for peer comm actors
(I think we have to *forward* in-order as well as deliver in-order, as we need
to know if a prev message was selected to be forwarded to subsequent nodes
to update this last-seq map, before forwaring subsequent messages). This
diff resolves this issue by packing all `RoutingFrame`s destined to a peer
comm actor into a single `ForwardMessage`.

Reviewed By: shayne-fletcher, moonli

Differential Revision: D73671610

fbshipit-source-id: dbf7c4f2fbfd77b8f51e1dfdc1bb32b1a36ecf14