Add support for reduce, and fix reduce bug
Summary:
This diff adds support for reduce in recordings. There was a bit of subtlety to deal with regarding error propagation where we have to explicitly check the input tensors for errors.

I also found a bug in the existing implementation of `reduce`. Previously, when `out` was defined for reductions, it was passed as the only tensor input to the reduction, with the contents of `local_tensor` being ignored -- so each rank would do the reduction with the original contents of `out` rather than the contents of `local_tensor`. This diff fixes the bug by copying the contents of `local_tensor` into `out` before running the reduction.

Reviewed By: colin2328

Differential Revision: D74113287

fbshipit-source-id: 18562b7edb85bcf0c6cc1b842824b6353aa10449