RemoteProcessAllocator cleanup on uplink breakage
Summary:
Add heartbeat to detect broken connection beteen `TaskGroupAlloc` and `RemoteProcessAllocator`. When detected, the `ProcessAlloc` is going to be stopped automatically.

I chose to make `heartbeat_interval` configurable by `TaskGroupAlloc` rather than locally parameterized because it is probably going to depend on the environment: larger worlds would requirer higher values so we don't flood the clients.

Note that because it relies on heartbeat, its frequency and its timeout, the exact time it would need to cleanly shutdown all processes may be different. During that time, if the clinet tries to reallocate it may fail. An additional enhancement will be needed to also reset the allocation if we get another Allocate request.

Reviewed By: mariusae

Differential Revision: D74603358

fbshipit-source-id: ec104019a3a2c4a7080bdf5b72f6665b570c8866