fix test_mock_cuda for oss
Summary: Synced up with Sam, the mock cuda setup is failing because cuda is already loaded on the process earlier in the testing sequence. This test should be done in a subprocess to isolate it, but I was running into some dependency errors. Instead for the time being we can run it separately after the main test suite.

Reviewed By: colin2328

Differential Revision: D74827312

fbshipit-source-id: 805946dd36da61dee78deba824fb4b4820ca79b9