Add bindings for NCCLX functions to nccl-sys
Summary:
Add bindings for a C++ API from NCCLX for `ncclCommDump`. This API takes in an
`unordered_map<string, string>` reference to add to, which is difficult to get working with cxx::bridge.
I made a small shim which instead appends pairs of strings to a vector which has easier bindings, then
in Rust turn that into a HashMap<String, String>.

This API will be useful for telemetry like CollTrace which is maintained by the NCCLX team.
The CollTrace fields will be written into keys like `CT_pastColls`, `CT_pendingColls`, and
`CT_currentColl`.

Needs to conditionally enable tests only when NCCLX is turned on, else the results won't apply.
So added a new feature that is available when NCCLX is on.

Also, there is not yet any connection from the controller to the worker for these APIs.

Differential Revision: D73476574