Add more logging to check_supervision in case it is failing
Summary:
For a OneWay message like `ControllerActor::check_supervision`, if it fails halfway through
it won't send a message to the client, and it won't reschedule itself to run again.
We need to have logs to make sure the message is sent.

I suspect that occasionally when a worker fails with a DeviceException it isn't getting propagated
back to the client

Reviewed By: shayne-fletcher

Differential Revision: D75232803

fbshipit-source-id: 25c6fa2e927a762305b8bbd73a181174a1129bda