mesh: propagate supervision events to the proc mesh
Summary:
This wires up the supervision mechanism to deliver supervision events to the proc mesh itself. Currently we interpret these as a proc crash.

This will then allow us to install a default policy to always fail on such errors (but, crucially, never silently). We can then drill into fine-grained delivery of the supervision events to the actor mesh level instead, from which they can be recovered.

Differential Revision: D74830956