mesh: use comm actor for casting
Summary:
This plumbs everything through so that `ActorMesh::cast` uses the comm actor. We always configure the comm actor in address book mode (since we perform our own rank assignment, and because we may break the mesh/world 1:1 relationship).

In order to propagate the rank through the envelope, we define a binding for a newtype, `CastRank` that is modified by the comm actor.

On the whole, the orchestration of castable methods (binding, cast envelope, registration, etc.) is become too complex, and we need to find a way to simplify this.

This still uses the root router. Next we'll introduce direct dialing.

Differential Revision: D75032514