eagerly import torch in actor bootstrap
Summary:
T225878580, there seems to be a race where under some conditions we don't fully initialize the torch module before attempting to access it. I suspect this has something to do with the different way that pickle imports globals vs. just normal import statements, but need to confirm.

I will try to root cause it further, but one thing that is certain to work is importing torch before entering into the actor runtime.

Reviewed By: mariusae

Differential Revision: D75645067