core configuration (#59)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/59

Replace ad-hoc configuration (read from environment variables) with a shared, global config. Provide utilities to modify the configuration temporarily for testing.

This change defines a shared, concrete `Config` struct. We'll probably want to replace this with an extensible API, so that we can apply the same configuration to hyperactor's various submodules.

Reviewed By: shayne-fletcher, moonli

Differential Revision: D75616316