Introduce non-advancing events (#151)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/151

TLDR;

We go from simulating 50 steps on a toy model D72348663 in
```
Simulated MNIST training completed in 341.3135597705841s
```
to
```
Simulated MNIST training completed in 7.598558664321899s
```
and can run fast enough to simulate llama3 and nanoGPT training in reasonable time

___________________________

Consider the following scenario with `self_message_with_delay`:

1. Event queue: [Sleep(10)], T = 0
2. Simnet dequeues Sleep(10) and advances to T=10
3. `self_message_with_delay` task wakes up and sends ControllerMessage::CheckSupervision
4. ControllerActor handles ControllerMessage::CheckSupervision and calls `self_message_with_delay(ControllerMessage::CheckSupervision)`
5. Event queue: [Sleep(10)], T = 20

This loops endlessly and we advance the time far further than we should.

Before we were remedying this in the Simnet event loop by adding a short debounce interval that would cause us to only move on to handling the events once the debounce duration has elapsed and no more events have arrived. In the above scenario this would mean that the events between T=10 and T=20 would arrive and be handled instead of instantly advancing to T=20, T=30, T=40...

The problem with this is this actually dramatically slows down the wall time of running the simulator such that examples like nanoGPT and llama3 would take very long to simulate a single step.

The solution to this is to have 2 separate queues:

queue 1: events that are allowed to advance the sim time on it's own
queue 2: events that are allowed to advance the sim time only when the earliest event in queue 1 comes after the earliest event in queue 2, or if some debounce period such that only queue 2 has events has elapsed. The second scenario occurs when there is a supervision failure

Doing this allows us to safely remove the debounce, dramatically speeding up the wall time it takes to run the simulator while also ensuring that events from `self_message_with_delay` are scheduled at the correct time

Reviewed By: kaiyuan-li

Differential Revision: D75900568