Add python accumulator and reducer
Summary:
We want to define accumulator and reducer in python, and run them in the Rust backend. This diff adds that.

Specifically, the implementation follows the patten established by `PythonActor`. The change consists of:

1. A rust `struct PythonAccumulator`, which is the container of 2 Python functions, one for accumulation and one for reduction.
2. A rust `PythonReducer`, which contains 1 Python function which is used for reduction.

The `struct PythonAccumulator` is bound to Python, and thus can take the Python functions in as parameters.

Differential Revision: D76156297