Dedicated Resource Manager, expose Queue Pairs (#272)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/272

This diff refactors the RDMA codebase to introduce the RdmaManagerActor as a dedicated resource manager for each endpoint, responsible for managing memory mappings and related communicators (QueuePairs).

Currently, the process is simple: we use CPU memory and expose the entire address space to the NIC, with a single QueuePair per route (i.e., all traffic between endpoints A and B goes through the same QP). As we move toward supporting GPUs, the requirements will become more complex. This refactor cleans up the API and abstractions in preparation for that transition.

In the new model, RdmaBuffers are extremely lightweight and should only be created by RdmaManagerActors. It's important to note that an RdmaBuffer is designed to carry only the minimum data necessary to the remote caller; under the hood, an RdmaBuffer represents a slice of a registered memory region with associated rkey and lkey.

Moving data between endpoints requires using the RdmaQueuePair's GET/PUT API. To obtain a QueuePair, it must be requested from an RdmaManagerActor.

core API for RdmaManagerActor:
     request_buffer(addr: u32, size: usize) -> RmdaBuffer.  // This acts a light-weight memory handle that can be serialized.  In practice this a slice of IBV MR maintained by RdmaManagerActor

     request_queue_pair(remote: ActorRef<RdmaManagerActor> -> RmdaQueuePair // light-weight wrapper around QP.

Example API usage:

  // setup envs, send remote_buffer -> local env.

  // establish RMDA connection with Peer if required.
  let qp = local_rmda_manager.request_queue_pair(client, remote_rmda_manager).await()?

  // get/put api example, returns immediately.
  qp.get(local_buffer, remote_buffer)?;

  // ... poll for completion

Differential Revision: D76370567