Fix panic in RemoteProcessAlloc select_all (#281)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/281

There was a bug in the remote process allocator, where if all hosts became closed, the loop
would hit a panic in `select_all` given an empty list of handles.
This change exits out of the loop instead, completing the task.

Added some asserts en route to the allocator to ensure it doesn't get a list of zero hosts to
query.

And some more `__repr__`'s which were helpful when debugging with logs.

Reviewed By: shayne-fletcher

Differential Revision: D76750206