Add `Castable` macro (#81)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/81

This diff adds a `Castable` derive marco. This macro implements `trait Bind` and `trait Unbind` for the applied struct/enum.

Right now this macro only supports the `PortRef` type. We could extend it to support arbitrary types once we have such need.

The expanded code is shown in this example: P1844273765

Differential Revision: D75322219