[7/n] move History into mesh_controller in prep for major refactor
mesh_controller will respond directly to PortId when things are finished inside controller. This is a big enough change that I am splitting the code from the other controller. This is just a wholesale copy of the history code, deleting the parts we don't use in this file. The next diff makes the changes needed for mesh_controller.

Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/277

Differential Revision: [D76643800](https://our.internmc.facebook.com/intern/diff/D76643800/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D76643800/)!
ghstack-source-id: 290789082