Castable ActorMeshRef (#328)
Summary:

We need to define and propagate actor mesh references, so that they can be passed and used to broadcast messages from other actors.

We will test this with a pingpong where the pingpong message contains an ActorMeshRef of the sender mesh, and we will use the `sel!(?)` to send to a single random actor on the other mesh.

```
ping:2 handling with ttl 10
pong:3 handling with ttl 9
ping:0 handling with ttl 8
pong:1 handling with ttl 7
ping:0 handling with ttl 6
pong:3 handling with ttl 5
ping:0 handling with ttl 4
pong:1 handling with ttl 3
ping:2 handling with ttl 2
pong:3 handling with ttl 1
ping:0 handling with ttl 0
```

Reviewed By: mariusae

Differential Revision: D76749054