Replace &Instance<A> with &Context<'_, A> in the Handler trait
Summary:
In order to make the headers accessible to message handlers in a cleaner way, this diff introduces a new struct `Context<'_, A>` which contains a `&Instance<A>` and the headers associated with the current message. As per T228176204, the `Handler` trait now takes a `&Context<Self>` for the `this` parameter. `Context` has the same capabilities as the contained `Instance`. Moreover, I have implemented `Deref` for `Context` to return `&Instance` -- with deref coercion, this means that no code has to change except for inside the method signatures (this deref coercion is also why the `Handler` takes `&Context` instead of `&Instance`.

Disclaimer: many of these changes were made by devmate.

Reviewed By: shayne-fletcher, mariusae

Differential Revision: D77341169