Added watchdog timeout when stop is issued for stuck actors. (#336)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/336

Allocators have a timeout to kill unresponsive procs when stop() is issued.

Here is the reason why we needed in the first place:

While running an xlformers test, we encountered a problem wherein the process was getting stuck upon issuing an exit.

The root cause for the blocking was the cuda unregistry routine getting stuck blocking the exit() call and all other calls.

In order to simulate the same, we added here an exit handler that loops forever.

Reviewed By: suo, technicianted

Differential Revision: D77303504