Track all `RootActorMesh`s spawned by `PyProcMesh` (#392)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/392

This diff adds a small wrapper around the `ProcMesh` wrapped by `PyProcMesh`,
to bookkeep all outstanding references to `RootActorMesh`s that get spawned (the
bookkeeping is automatically GC'd when outstanding references are dropped).

The spawned `RootActorMesh`s are also now wrapped by `SharedCell` to allow
the `ProcMesh` to dispose of them on close, and to require callers to perform a
fallible "borrow" on access, which is used by subsequent diffs to allow for a graceful
cleanup/drop of a `ProcMesh` while still allowing outstanding-but-invalid references
in Python code.

Reviewed By: mariusae

Differential Revision: D77385189