Add a preemptible `RWLock` utility (#409)
Summary:

This adds a `tokio::sync::RWLock` wrapper struct which allows
waiting writers to "preempt" and interrupt readers that hold the
lock.

This is supported by a `async fn preempt(&self)` helper on the 
read guard that readers can use in a `tokio::select!` to poll for
interruptions and vacate their lock guard.

Reviewed By: suo

Differential Revision: D77635672