No more operational messages (#473)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/473

Operational messages were used in order to signal to the simulator that it should perform certain actions like growing or shrinking the mesh. This was needed since the python and rust were running in separate processes, and messages were needed to communicate between the two, but now everything is on the same process so we can do this in memory.

Differential Revision: D77941643