fix pid in log_writer
Summary:
Previously process ID was derived wrong by using the process ID of the current process. That is wrong because the log_writer is spawned on the parent process.
This diff corrects it by attaching the Child process's pid to the log writer so this pid can be used directly in generating the Log Object Name.

Reviewed By: highker

Differential Revision: D78105348