(2/N) Add in monarch/python/_src/tensor_engine (#486)
Summary:

This diff adds in a dedicated target for tensor engine.

In rdma.py, adds in basic support for RDMABuffer CPU.
While CUDA<>CPU|CUDA is implemented in monarch_rdma, this was running into issues with MR registration.
Logging shows that the code path correctly interprets the address as CUDA, but cuMemGetHandleForAddressRange returns the handle/fd as -1 and a null pointer MR.
RDMABuffer will only support CPU for now with GPU support as a follow up, cc dstaay-fb

Reviewed By: mariusae

Differential Revision: D78035395