Push MeshTrait implementation to _ActorMeshRefImpl
Summary:
`_ActorMeshRefImpl` is the standin class for `PythonActorMesh` and `PythonActorMeshRef`. This diff stack is working on replacing `_ActorMeshRefImpl` with `PythonActorMesh` and `PythonActorMeshRef`.

Compared to `PythonActorMesh` and `PythonActorMeshRef`, one method `_ActorMeshRefImpl` is missing is the `def slice` method.  Lacking of this method blocks to do a drop-in replacement.

This diff pushes the `MeshTrait` implementation to `_ActorMeshRefImpl`. In this way, `_ActorMeshRefImpl` will have `the slice` method from `MeshTrait`.

Differential Revision: D78300586