Lift the `send-to-actor` logic from `_ActorMeshRefImpl` to `ActorIdRef` (#555)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/555

`RDMAManger` and `DebugManager` have use cases that need to send messages directly to an actor based on its actor id, instead of to a mesh. Currently the "send-to-actor" function is piggybacking `_ActorMeshRefImpl`'s implementation. This is blocking the migration from `_ActorMeshRefImpl` to `PythonActorMesh`, because we cannot create a mesh solely based on its actor ID.

To unblock, this diff lifts the `send-to-actor` logic from  `_ActorMeshRefImpl` to `ActorIdRef`, so it is decoupled from `_ActorMeshRefImpl` and `ActorMeshRef`.

Differential Revision: D78302352