Flush ack before ServerConn.process exits (#573)
Summary:

Currently when `ServerConn.process` exits, it does not flush the pending ack. This is okay for normal scenarios, because the connection will be reestablished by `NetTx`, and the ack will be sent in the new connection. However, this is problematic during shutdown, when the server side will be completely gone. In that case, if `NetTx` is still around, it will error out due to ack timeout.

This diff adds the flush which should fix this problem.

Reviewed By: vidhyav

Differential Revision: D78514364