Spawn signal handler out of loop into background task
Summary:
This fixes `test_allocator_stuck_job`.

Previously we had a problem where if we branched into any of the message handler branches where we called `std::process::exit()` the loop would break and we would not be able to receive any signals. This would result in us being unable to kill the process with SIGTERM which is why the broken test would hang. Spawning the signal handler in a background task ensures that even once the loop breaks we can still listen for signals and handle them

Differential Revision: D78516491