Add timeout for connection from client to remote process allocator (#583)
Summary:

Adds a `--timeout` option to RemoteAllocator. This will define the time the allocator will wait
for a connection from the client before exiting. The timeout will not hit if it keeps receiving
messages or has an active allocation.

Current default is 0 (infinite) to keep parity with existing uses. This can be changed in
the future.

Reviewed By: pablorfb-meta

Differential Revision: D78567664