at least show something when init fails
Summary: The goal is to show why `__init__` fails when creating a actor mesh. This is not a proper fix as the right solution should follow T229200522. I was halfway into the issue until I found T229200522. So pzhan9, I will leave it to you to get the right fix. This diff is a mitigation.

Differential Revision: D78606675