Use async functions in sync context to avoid code duplication. (#587)
Summary:

We currently have a lot of code duplication for blocking/nonblocking variants.

This PR lets the blocking variants be defined in terms of the async ones.

It first directly exposes the tokio Future object to python, letting it choose to synchronously block on it, or turn it into a asyncio.Future. Then when we are running async code to do pipe receives in a sync context, we can skip running a real event loop and just step the coroutine manually.

See [avoiding async code duplication] in the diff for a description of the approach.


Currently there is jank around the fact that we are running synchronous actor code on a thread that already has a running event loop. To indicate to the synchronous actor that it is ok to `get()` despite this setup, we temporarily blank out the running asyncio loop while a asynchronous actor is running a message. We need to eventually figure out a way to get the synchronous actors running on a thread that truly is not the asyncio event loop.

Reviewed By: eliothedeman, colin2328

Differential Revision: D78466520