Stop instead of terminate RemoteProcessAllocator on SIGINT
Summary:
```
this is not the desired behavior. After this diff, if I send ctrl-C to my client process, the signal handlers will send a termination message to the RemoteProcessAllocator, resulting in this loop being broken out of and my MAST job exiting.
This prevents me from interactively using my MAST job. I think the right behavior would be to just stop the current alloc on a signal failure, not exit the RemoteProcessAllocator loop entirely (e.g. send a RemoteProcessAllocatorMessage::STOP instead of this new TEMRINATE.
```

Reviewed By: vidhyav

Differential Revision: D78897304